Add UTC Timezone Methods
By default the go transit API returns data in the America/Toronto
timezone. More often that not dates are stored in databases as UTC
timezone, I want to add helper methods that will take care of the
conversion so no additional work is required. Add `_utc` to any date and
a UTC date object will be returned.

This change addresses the need by:
* Add utc date helpers to data points using a date/time