Fix Failing Specs
Specs are failing and it appears to have 2 root causes. 1. github
checkout has been upgraded to v4 and the v3 in use is no longer
available. 2. sinatra, used in the test suite, has an added layer of
security that needs to be disabled for tests to be successful.

This change addresses the need by:
* Update action checkout to v4
* Add bypass for allowed hosts for fake metrolinx
* Remove support for ruby < 2.7.0